server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langchain>=1.3.1",
    "langchain-openai>=1.2.1",
    "langgraph>=1.2.0",
    "langsmith>=0.8.5",
    "pydantic>=2.13.4",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.47.0",
]
